Render try_read events and extend "should_resolve" threads
If a thread attempts to read from another one, but it never completes,
show a yellow arrow at the attempt so it's clear why it didn't complete.

If a thread is marked as "should_resolve" then mark it as failed if it
never resolved and set its end time to the end of the trace to make it
more visible.